test(task-migrate-clean-batch): 迁 gui_lifecycle 的 7 个无帧用例到 gui_unit_test（清单 9→7，见下）
420.5 Step 10。部分迁移 7/12。

**与 420.3 清单「9/12」的偏差：实际可迁 7 个，原因已定位**
420.3 用「函数体开头有 IM_UNUSED(ctx)」当匿名 ctx 的静态判据，而这条判据在本文件
产生两个假阳性：run_completed_latch_survives_invalid_snapshot 与
display_edits_do_not_disturb_done 开头确实写了 IM_UNUSED(ctx)，函数体后半段却仍
用 ctx->Yield() 当帧泵。按权威判据（不用 ImGuiTestContext）它们不可迁，留在 gui_test。
偏差 2 个，落在 plan §7 风险 4 声明的容差内，不为凑数字放宽判据。

**保留对真实全局的操作是忠实迁移，不是偷懒**
这 7 个直接读写 gui::g_state / gui::g_server_poller / gui::SyncFromPoller()。
SyncFromPoller() 与 ReconcileSimState() 的调用方都不收 state 参数——「单一 owner
每帧 reconcile 真实全局」本身就是被测的设计，注入局部 GuiState 等于去测一条生产里
不存在的路径（plan §3「去全局化的适用边界」）。每个用例只重置自己碰的那几个字段。

- 新增 test/unit-correctness/gui/test_lifecycle.cpp（7 个 TEST + 迁走的 helper 簇）
- 原文件删 7 个块（12→5），头注释重写为"这里只剩要活帧的五个"
- kFiniteConfig / CommitJsonConfig / RunFiniteToCompletion / CurrentEpoch 四个 helper
  在原文件已无消费者，随之删除（该 target 没开 -Wunused-function，不删就是静默死代码）

核验：源文件侧断言减量 IM_CHECK 47 + IM_CHECK_EQ 44 = 91，新文件
EXPECT_TRUE 47 + EXPECT_EQ 44 = 91，按宏形态逐类对上；
gui_test 静态用例数 498→491；gui_unit_test 77→84 全绿；
**--gtest_shuffle --gtest_repeat=3 三轮 84/84 全绿**——这批共享真实全局单例，
顺序敏感是本步最大风险点，已按 plan Step 10 测试点实测排除。
